Simplify texture disposal logic
