+++ osu.Framework/Graphics/Textures/TextureWithRefCount.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
+using System;
 using osu.Framework.Graphics.OpenGL.Textures;
 using OpenTK.Graphics.ES30;
 
@@ -9,7 +10,7 @@ namespace osu.Framework.Graphics.Textures
     /// <summary>
     /// A texture which updates the reference count of the underlying <see cref="TextureGL"/> on ctor and disposal.
     /// </summary>
-    public class TextureWithRefCount : Texture
+    public class TextureWithRefCount : Texture, IDisposable
     {
         public TextureWithRefCount(TextureGL textureGl)
             : base(textureGl)
@@ -23,12 +24,29 @@ namespace osu.Framework.Graphics.Textures
             TextureGL.Reference();
         }
 
-        protected override void Dispose(bool isDisposing)
-        {
-            if (!IsDisposed)
-                TextureGL?.Dereference();
+        #region Disposal
 
-            base.Dispose(isDisposing);
+        public bool IsDisposed { get; private set; }
+
+        ~TextureWithRefCount()
+        {
+            Dispose(false);
         }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool isDisposing)
+        {
+            if (IsDisposed)
+                return;
+            IsDisposed = true;
+            TextureGL?.Dereference();
+        }
+
+        #endregion
     }
 }
