+++ osu.Framework/Graphics/UserInterface/CircularProgress.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
 
+using System;
 using osu.Framework.Allocation;
 using osu.Framework.Configuration;
 using osu.Framework.Graphics.Shaders;
@@ -30,11 +31,8 @@ namespace osu.Framework.Graphics.UserInterface
 
         protected override void Dispose(bool isDisposing)
         {
-            if (CanDisposeTexture)
-            {
-                texture?.Dispose();
-                texture = null;
-            }
+            (texture as IDisposable)?.Dispose();
+            texture = null;
 
             base.Dispose(isDisposing);
         }
@@ -80,8 +78,7 @@ namespace osu.Framework.Graphics.UserInterface
                 if (value == texture)
                     return;
 
-                if (texture != null && CanDisposeTexture)
-                    texture.Dispose();
+                (texture as IDisposable)?.Dispose();
 
                 texture = value;
                 Invalidate(Invalidation.DrawNode);
